+++ Changelog.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 3.x
+
+### Changed
+
+  * Refactored and renamed method `ftp-sync` to `artifacts--ftp` in preparation of git-sync.
+ 
+### New
+  * `rootFolder` is set by default now
+  * All context variables are exposed as replacement patterns for using in scripts.
+  * new method `artifacts--git` to build an artifact and push it to a git repository, see new documentation about artifacts.
+
 ## 3.1.0 / 2019-09-27
 
 ### New